requirements.txt
amqp==5.1.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.26.51
botocore==1.29.51
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
cryptography==39.0.0
defusedxml==0.7.1
Deprecated==1.2.13
django-allauth==0.52.0
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap-modal-forms==2.2.0
django-bootstrap5==22.1
django-celery-beat==2.4.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-extensions==3.2.1
django-filter==22.1
django-model-utils==4.2.0
django-storages==1.12.3
django-tables2==2.4.1
django-timezone-field==5.0
django-widget-tweaks==1.4.2
idna==3.4
jmespath==0.10.0
kombu==5.2.4
oauthlib==3.2.2
packaging==23.0
Pillow==9.2.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.6.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
wrapt==1.14.1
